添加导入密码功能

+++ app/src/main/java/com/xiaomai/geek/ui/module/password/PasswordSettingFragment.java
@@ -3,11 +3,11 @@ package com.xiaomai.geek.ui.module.password;
 
 import android.app.AlertDialog;
 import android.app.ProgressDialog;
-import android.content.DialogInterface;
 import android.os.Bundle;
 import android.support.annotation.Nullable;
 import android.support.design.widget.Snackbar;
 import android.support.design.widget.TextInputLayout;
+import android.text.TextUtils;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -43,6 +43,9 @@ public class PasswordSettingFragment extends BaseFragment implements IPasswordSe
     @BindView(R.id.layout_modify_password)
     LinearLayout layoutModifyPassword;
 
+    @BindView(R.id.layout_import)
+    LinearLayout layoutImport;
+
     private PasswordSettingPresenter mPresenter = new PasswordSettingPresenter();
 
     public static PasswordSettingFragment newInstance() {
@@ -65,7 +68,8 @@ public class PasswordSettingFragment extends BaseFragment implements IPasswordSe
     }
 
     @OnClick({
-            R.id.layout_clear_data, R.id.layout_backup, R.id.layout_modify_password
+            R.id.layout_clear_data, R.id.layout_backup, R.id.layout_modify_password,
+            R.id.layout_import
     })
     public void onClick(View view) {
         switch (view.getId()) {
@@ -78,38 +82,44 @@ public class PasswordSettingFragment extends BaseFragment implements IPasswordSe
             case R.id.layout_modify_password:
                 showModifyDialog();
                 break;
+            case R.id.layout_import:
+                mPresenter.importPassword(mContext);
+                break;
         }
     }
 
-    private EditTextDialog mDialog;
-
     private void showModifyDialog() {
-        mDialog = new EditTextDialog.Builder(mContext).setTitle("修改密码")
-                .setHint("请输入新密码")
+        new EditTextDialog.Builder(mContext).setTitle("修改密码").setHint("请输入新密码")
                 .setOnPositiveButtonClickListener(
                         new EditTextDialog.Builder.OnPositiveButtonClickListener() {
                             @Override
-                            public void onClick(TextInputLayout textInputLayout, String password) {
+                            public void onClick(EditTextDialog dialog,
+                                    TextInputLayout textInputLayout, String password) {
                                 if (password.length() < 6) {
                                     textInputLayout.setError("密码长度不能小于6");
                                 } else {
                                     PasswordPref.savePassword(mContext, password);
-                                    mDialog.dismiss();
+                                    dialog.dismiss();
                                     Snackbar.make(layoutBackup, "密码修改成功！", Snackbar.LENGTH_LONG)
                                             .show();
                                 }
                             }
                         })
-                .create();
-        mDialog.show();
+                .create().show();
     }
 
     private void clearData() {
-        new AlertDialog.Builder(mContext).setMessage("您确定要清空所有数据吗？").setNegativeButton("取消", null)
-                .setPositiveButton("确定", new DialogInterface.OnClickListener() {
+        new EditTextDialog.Builder(mContext).setTitle("清空数据").setOnPositiveButtonClickListener(
+                new EditTextDialog.Builder.OnPositiveButtonClickListener() {
                     @Override
-                    public void onClick(DialogInterface dialog, int which) {
-                        mPresenter.deleteAllPasswords(mContext);
+                    public void onClick(EditTextDialog dialog, TextInputLayout textInputLayout,
+                            String password) {
+                        if (TextUtils.equals(PasswordPref.getPassword(mContext), password)) {
+                            mPresenter.deleteAllPasswords(mContext);
+                            dialog.dismiss();
+                        } else {
+                            textInputLayout.setError("密码错误，请重试！");
+                        }
                     }
                 }).create().show();
     }
@@ -134,15 +144,29 @@ public class PasswordSettingFragment extends BaseFragment implements IPasswordSe
     public void onBackupComplete(int count) {
         if (mProgressDialog != null && mProgressDialog.isShowing())
             mProgressDialog.dismiss();
-        Snackbar.make(layoutBackup, "成功备份" + count + "条数据", Snackbar.LENGTH_LONG).show();
+        if (count > 0) {
+            Snackbar.make(layoutBackup, "成功备份" + count + "条数据", Snackbar.LENGTH_LONG).show();
+        } else {
+            Snackbar.make(layoutBackup, "没有备份任何数据", Snackbar.LENGTH_LONG).show();
+        }
     }
 
     private AlertDialog mProgressDialog;
-
 
     @Override
     public void showBackupIng() {
         mProgressDialog = new ProgressDialog.Builder(mContext).setMessage("正在备份中").create();
         mProgressDialog.show();
+    }
+
+    @Override
+    public void importComplete(int count) {
+        Snackbar.make(layoutBackup, "成功导入" + count + "条数据", Snackbar.LENGTH_LONG).show();
+        EventBus.getDefault().post(new PasswordEvent(PasswordEvent.TYPE_ADD, new Password()));
+    }
+
+    @Override
+    public void importFail(String message) {
+        Snackbar.make(layoutBackup, "导入失败，请检查文件", Snackbar.LENGTH_LONG).show();
     }
 }
